+++ Assets/RoDOTS/Runtime/Controllers/GameController.cs
@@ -1,7 +1,0 @@
-﻿using EugeneC.Singleton;
-
-namespace RoDOTS.runtime {
-
-    public class GameController : GenericSingleton<GameController> { }
-
-}