+++ Assets/RoDOTS/Runtime/Game/GameController.cs
@@ -1,0 +1,9 @@
+﻿using EugeneC.Singleton;
+
+namespace RoDOTS.runtime {
+
+    public class GameController : GenericSingleton<GameController> {
+        
+    }
+
+}