+++ Assets/RoDOTS/Runtime/UI/BasicUI.cs
@@ -1,0 +1,27 @@
+﻿using EugeneC.Mono;
+
+namespace RoDOTS.runtime {
+    
+    public class BasicUI : UiHelper {
+        
+        public override void OnSpawn() {
+            
+        }
+        
+        public override float OnStartOpen() {
+            return transitionTime;
+        }
+        
+        public override void OnEndOpen() {
+            
+        }
+        
+        public override float OnStartClose() {
+            return transitionTime;
+        }
+        
+        public override void OnEndClose() {
+            
+        }
+    }
+}